refactor(query): Split mapper/filter/filter_impl.rs into submodules
